requirements.txt
nltk==3.8.1
gensim==4.3.2
numpy==1.26.0
streamlit==1.28.1 
transformers==4.35.2
matplotlib==3.8.1
torch==2.1.1
pillow==10.1.0
